Bug 1091643 - Switch bugscache queries from full-text search to LIKE
Since full-text searches miss cases where the search term was truncated,
but is still a substring of the bug summary.

When using LIKE, any underscores or percent signs will be treated as
wildcards, so we need to escape them, as well as the escape symbol
itself, so the search term is handled literally. A custom escape
symbol is used since the default of backslash leads to less readable
code, given it has to be double escaped.