Bug 1078239 - Fix bug suggestions for recently modified, resolved bugs
We divide bug suggestions for a search term into 'open_recent' and
'all_others'. The former is supposed to be group #1 below, and the
latter groups 2-4, with the two groups combined corresponding to every
bug whose summary matches the search term.

1) Open + recently modified
2) Open + not recently modified
3) Resolved + recently modified
4) Resolved + not recently modified

However prior to this patch group #3 was not being returned at all, when
it should have been included in all_others.